Split id generator into individual class files
