+++ ruby/lib/cucumber/messages/id_generator/incrementing.rb
@@ -1,0 +1,16 @@
+module Cucumber
+  module Messages
+    module IdGenerator
+      class Incrementing
+        def initialize
+          @index = -1
+        end
+
+        def new_id
+          @index += 1
+          @index.to_s
+        end
+      end
+    end
+  end
+end
