+++ ruby/lib/cucumber/messages/id_generator/uuid.rb
@@ -1,0 +1,13 @@
+require 'securerandom'
+
+module Cucumber
+  module Messages
+    module IdGenerator
+      class UUID
+        def new_id
+          SecureRandom.uuid
+        end
+      end
+    end
+  end
+end
